refactor(cli-core): strip remaining cfg(not(wasm32)) gates from leo-package
Move every native-only piece left in `crates/leo-package` over to
`leo-cli-core`:

- `is_valid_program_name` / `is_valid_library_name` / `reserved_keywords`
  / `verify_valid_program` → `leo_cli_core::validation::*`. The bodies
  pull in snarkVM's keyword tables.
- `fetch_from_network` / `fetch_from_network_plain` /
  `fetch_program_from_network` / `fetch_latest_edition` /
  `create_http_agent` / `retry_network_call` → `leo_cli_core::network::*`.
  These used `ureq` (HTTP).
- `MAX_PROGRAM_SIZE` compile-time assertion against snarkVM's constant
  → `leo_cli_core::validation` (kept; the literal itself stays in
  `leo-package` where every dep walker reads it).
- Native-only validation branch of `parse_dependencies_from_aleo` is
  gone — the inline scanner (already wasm-buildable) is now the
  canonical extractor on every target. Callers that want strict
  bytecode validation call `verify_valid_program` separately.
- Snarkvm parity test for the scanner moved to
  `leo_cli_core::validation::tests`. `parse_dependencies_from_aleo`'s
  helper `extract_aleo_import_names` becomes `pub`.
- `Workspace` module is no longer cfg-gated. The
  `is_valid_library_name` call in `Workspace::initialize_skeleton`
  moves to the CLI's `LeoNew` handler (the only caller). The
  workspace-dep resolution `resolve_workspace_dependency_with_file_source`
  drops its native-only branch.
- `snarkvm` and `ureq` are no longer in `crates/leo-package`'s native
  deps (only `leo_cli_core` uses them now).

Net: `crates/leo-package` has *zero* `#[cfg(...)]` gates and is purely
wasm-buildable. Same for `crates/leo/src/lib.rs` (which was already
clean after earlier commits). The CLI keeps depending on `leo-cli-core`
for what was native-only, and `leo-wasm` continues to pull only the
wasm-buildable surface.

`crates/leo`'s CLI commands switch their `leo_package::fetch_*` /
`leo_package::is_valid_*` / `leo_package::verify_valid_program` imports
to `leo_cli_core::network::*` and `leo_cli_core::validation::*`. The
benchmarks crate passes `leo_package::reject_network_fetcher` (a new
helper) for its local-only fixture builds since it doesn't have a
network fetcher to wire in.
