+++ crates/leo-cli-core/Cargo.toml
@@ -38,6 +38,12 @@ indexmap   = { workspace = true }
 serde      = { workspace = true }
 serde_json = { workspace = true }
 
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+# `network`, `validation`, `package_fetch`, `package_init` modules use these.
+snarkvm = { workspace = true }
+tracing = { workspace = true }
+ureq    = { workspace = true }
+
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 # `clap` (and friends) pull `getrandom 0.4` in transitively on wasm32; that
 # crate refuses to compile unless the `wasm_js` backend is opted in.
