+++ crates/leo/src/cli/commands/query/mod.rs
@@ -15,7 +15,7 @@
 // along with the Leo library. If not, see <https://www.gnu.org/licenses/>.
 
 use leo_ast::NetworkName;
-use leo_package::{fetch_from_network, verify_valid_program};
+use leo_cli_core::{network::fetch_from_network, validation::verify_valid_program};
 
 use super::*;
